> moved command handler as per #23
