+++ handlers/message.js
@@ -1,5 +1,5 @@
 const CONFIG = require("../modules/config");
-const commandHandler = require("./command").handler;
+const commandHandler = require("./command/commandHandler").handler;
 const handleDataCheck = require("../checks/handleData").check;
 const dbInt = require("../db/interface");
